refactor(home): move to react
