+++ app/react/components/EdgeIndicator.test.tsx
@@ -12,12 +12,12 @@ test('when edge id is not set, should show unassociated label', async () => {
   expect(unassociatedLabel).toBeVisible();
 });
 
-// test('given edge id and last checkin is set, should show heartbeat', async () => {
-//   const { queryByLabelText } = await renderComponent('id', 1);
+test('given edge id and last checkin is set, should show heartbeat', async () => {
+  const { queryByLabelText } = await renderComponent('id', 1);
 
-//   expect(queryByLabelText('edge-heartbeat')).toBeVisible();
-//   expect(queryByLabelText('edge-last-checkin')).toBeVisible();
-// });
+  expect(queryByLabelText('edge-heartbeat')).toBeVisible();
+  expect(queryByLabelText('edge-last-checkin')).toBeVisible();
+});
 
 async function renderComponent(
   edgeId = '',
