+++ app/react/components/EdgeIndicator.tsx
