+++ app/react/portainer/HomeView/EnvironmentList/EnvironmentItem/EnvironmentIcon.tsx
