+++ app/react/portainer/HomeView/EnvironmentList/EnvironmentItem/EnvironmentStats.tsx
