+++ app/react/portainer/HomeView/EnvironmentList/EnvironmentItem/index.ts
@@ -1,2 +1,1 @@
 export { EnvironmentItem } from './EnvironmentItem';
-export { EdgeIndicator } from './EdgeIndicator';
