+++ app/react/portainer/HomeView/EnvironmentList/EnvironmentList.tsx
@@ -12,10 +12,6 @@ import {
   EdgeTypes,
 } from '@/react/portainer/environments/types';
 import { EnvironmentGroupId } from '@/portainer/environment-groups/types';
-import {
-  HomepageFilter,
-  useHomePageFilter,
-} from '@/portainer/home/HomepageFilter';
 import { useDebounce } from '@/portainer/hooks/useDebounce';
 import {
   refetchIfAnyOffline,
@@ -23,14 +19,12 @@ import {
 } from '@/react/portainer/environments/queries/useEnvironmentList';
 import { useGroups } from '@/portainer/environment-groups/queries';
 import { useTags } from '@/portainer/tags/queries';
-import { Filter } from '@/portainer/home/types';
 import { useAgentVersionsList } from '@/react/portainer/environments/queries/useAgentVersionsList';
 import { EnvironmentsQueryParams } from '@/react/portainer/environments/environment.service';
 import { useUser } from '@/portainer/hooks/useUser';
 
 import { TableFooter } from '@@/datatables/TableFooter';
 import { TableActions, TableContainer, TableTitle } from '@@/datatables';
-import { SortbySelector } from '@@/datatables/SortbySelector';
 import {
   FilterSearchBar,
   useSearchBarState,
@@ -38,6 +32,9 @@ import {
 import { Button } from '@@/buttons';
 import { PaginationControls } from '@@/PaginationControls';
 
+import { SortbySelector } from './SortbySelector';
+import { HomepageFilter, useHomePageFilter } from './HomepageFilter';
+import { Filter } from './types';
 import { EnvironmentItem } from './EnvironmentItem';
 import { KubeconfigButton } from './KubeconfigButton';
 import { NoEnvironmentsInfoPanel } from './NoEnvironmentsInfoPanel';
