+++ app/react/portainer/HomeView/EnvironmentList/KubeconfigButton/KubeconfigButton.tsx
