+++ app/react/portainer/HomeView/EnvironmentList/KubeconfigButton/index.ts
