+++ app/react/portainer/HomeView/EnvironmentList/NoEnvironmentsInfoPanel.tsx
