+++ app/react/portainer/HomeView/EnvironmentList/types.ts
@@ -1,0 +1,4 @@
+export interface Filter<T = number> {
+  value: T;
+  label: string;
+}
