+++ app/react/portainer/HomeView/LicenseNodePanel.test.tsx
@@ -1,7 +1,6 @@
 import { server, rest } from '@/setup-tests/server';
 import { renderWithQueryClient } from '@/react-tools/test-utils';
-
-import { LicenseType } from '../license-management/types';
+import { LicenseType } from '@/portainer/license-management/types';
 
 import { LicenseNodePanel } from './LicenseNodePanel';
