+++ app/react/portainer/HomeView/home.service.ts
@@ -1,4 +1,4 @@
-import axios, { parseAxiosError } from '../services/axios';
+import axios, { parseAxiosError } from '@/portainer/services/axios';
 
 import { Motd } from './types';
