+++ app/react/portainer/HomeView/types.ts
@@ -5,8 +5,3 @@ export interface Motd {
   Style?: string;
   ContentLayout?: Record<string, string>;
 }
-
-export interface Filter<T = number> {
-  value: T;
-  label: string;
-}
